changed the rowbytes, width, height, conwidth and conheight members of viddef_t from unsigned to signed int. fixed the -Wsign-compare warnings for NUM_GLMODES. fixed compilation using g++, by adding a missing type cast.
git-svn-id: svn://svn.code.sf.net/p/quakespasm/code/trunk/quakespasm@641 af15c1b1-3010-417e-b628-4374ebc0bcbd

+++ Quake/keys.c
@@ -274,7 +274,7 @@ void Key_Console (int key)
 	// when mouse is released to the window manager
 	// case K_MWHEELUP:
 		con_backscroll += keydown[K_CTRL] ? ((con_vislines>>3) - 4) : 2;
-		if (con_backscroll > con_totallines - (int)(vid.height>>3) - 1)
+		if (con_backscroll > con_totallines - (vid.height>>3) - 1)
 			con_backscroll = con_totallines - (vid.height>>3) - 1;
 		return;
